+++ src/components/ExchangeCalc.tsx
@@ -1,0 +1,7 @@
+export default function ExchangeCalc() {
+  return (
+    <div>
+      <input /> 유로 ▶︎ <input disabled /> 원
+    </div>
+  );
+}
